moving the actions out the the views
