+++ redux-intro/src/todo/components/TodoList.tsx
@@ -4,6 +4,7 @@ import TodoComponent from './Todo';
 import { TodoAppState } from '../..';
 import { TodoAction } from '../todoReducer';
 import { connect } from 'react-redux';
+import { toggleTodo } from './actionCreators';
 
 export interface TodoListProps {
   todos: Todo[];
@@ -38,10 +39,7 @@ const mapStateToProps = (state: TodoAppState) => {
 const mapDispatchToProps = (dispatch: (action: TodoAction) => void) => {
   return {
     onTodoClick: (id: number) => {
-      dispatch({
-        id,
-        type: 'TOGGLE_TODO'
-      });
+      dispatch(toggleTodo(id));
     }
   };
 };
